Update BlueForm to HTML5 and add 'arrangements'.
Created a new "tag" method which creates all the tags. The earlier methods, such as input_text, et.al., now just call "tag". Also added an "arrangement" option to allow the label-control fields to be wrapped in paragraphs (the default), or a table for nicer formatting, or no wrapping at all.